Fixed gs.engine.R so that the vignette works.
**NOTE** My fix is as follows:
The old code was failing because the example with the simple tables was broken.

It was failing because the code reads four lines, an d then was expecting to read the table data. However, the fifth line contains the table header

My fix is to read the table header line (and process it into a vector of length 4), and to delete the line where the header was set manually.

I'm not convinced that this is robust. I need to ask David whether I can always expect just four lines and then a header when I read from a spreadsheet. It seems unlikely.